Moving config migration tests out of bg_utils
