+++ src/app/test/config_test.py
@@ -5,7 +5,7 @@ from pathlib import Path
 import pytest
 import yapconf
 from box import Box
-from mock import Mock
+from mock import Mock, patch
 from ruamel import yaml
 from yapconf import YapconfSpec
 
@@ -172,6 +172,189 @@ class TestConfigGet(object):
         assert beer_garden.config.get("amq.host") == "localhost"
         assert beer_garden.config.get("INVALID_KEY") is None
         assert beer_garden.config.get("") is None
+
+
+class TestSafeMigrate(object):
+    @pytest.fixture
+    def spec(self, monkeypatch):
+        # return YapconfSpec(
+        raw_spec = {
+            "log": {
+                "type": "dict",
+                "items": {
+                    "config_file": {
+                        "type": "str",
+                        "description": "Path to a logging config file.",
+                        "required": False,
+                        "cli_short_name": "l",
+                        "previous_names": ["log_config"],
+                        "alt_env_names": ["LOG_CONFIG"],
+                    },
+                    "file": {
+                        "type": "str",
+                        "description": "File you would like the application to log to",
+                        "required": False,
+                        "previous_names": ["log_file"],
+                    },
+                    "level": {
+                        "type": "str",
+                        "description": "Log level for the application",
+                        "default": "INFO",
+                        "choices": [
+                            "DEBUG",
+                            "INFO",
+                            "WARN",
+                            "WARNING",
+                            "ERROR",
+                            "CRITICAL",
+                        ],
+                        "previous_names": ["log_level"],
+                    },
+                },
+            },
+            "configuration": {
+                "type": "dict",
+                "bootstrap": True,
+                "items": {
+                    "file": {
+                        "required": False,
+                        "bootstrap": True,
+                        "cli_short_name": "c",
+                    },
+                    "type": {
+                        "required": False,
+                        "bootstrap": True,
+                        "cli_short_name": "t",
+                    },
+                },
+            },
+        }
+        # )
+
+        monkeypatch.setattr(beer_garden.config, "_SPECIFICATION", raw_spec)
+
+        return raw_spec
+
+    @pytest.fixture
+    def old_config(self):
+        """Represent an un-migrated config with previous default values."""
+        return {
+            "log_config": None,
+            "log_file": None,
+            "log_level": "INFO",
+            "configuration": {"type": "json"},
+        }
+
+    @pytest.fixture
+    def new_config(self):
+        """Represents a up-to-date config with all new values."""
+        return {"log": {"config_file": None, "file": None, "level": "INFO"}}
+
+    @pytest.fixture
+    def config_file(self, tmpdir):
+        return Path(tmpdir, "config.yaml")
+
+    def test_success(self, tmpdir, spec, config_file, old_config, new_config):
+        old_config["configuration"]["file"] = str(config_file)
+
+        with open(config_file, "w") as f:
+            yaml.safe_dump(
+                old_config, stream=f, default_flow_style=False, encoding="utf-8"
+            )
+
+        generated_config = beer_garden.config.load(["-c", str(config_file)], force=True)
+        assert generated_config.log.level == "INFO"
+
+        with open(config_file) as f:
+            new_config_value = yaml.safe_load(f)
+
+        assert new_config_value == new_config
+        assert len(os.listdir(tmpdir)) == 2
+
+    def test_no_change(self, tmpdir, spec, config_file, new_config):
+        with open(config_file, "w") as f:
+            yaml.safe_dump(
+                new_config, stream=f, default_flow_style=False, encoding="utf-8"
+            )
+
+        generated_config = beer_garden.config.load(["-c", str(config_file)], force=True)
+        assert generated_config.log.level == "INFO"
+
+        with open(config_file) as f:
+            new_config_value = yaml.safe_load(f)
+
+        assert new_config_value == new_config
+        assert len(os.listdir(tmpdir)) == 1
+
+    def test_migration_failure(
+        self, monkeypatch, capsys, tmpdir, spec, config_file, old_config
+    ):
+        monkeypatch.setattr(
+            beer_garden.config.YapconfSpec,
+            "migrate_config_file",
+            Mock(side_effect=ValueError),
+        )
+
+        with open(config_file, "w") as f:
+            yaml.safe_dump(
+                old_config, stream=f, default_flow_style=False, encoding="utf-8"
+            )
+
+        generated_config = beer_garden.config.load(["-c", str(config_file)], force=True)
+
+        # Make sure we printed something
+        assert capsys.readouterr().err
+
+        # If the migration fails, we should still have a single unchanged JSON file.
+        assert len(os.listdir(tmpdir)) == 1
+        with open(config_file) as f:
+            new_config_value = yaml.safe_load(f)
+        assert new_config_value == old_config
+
+        # And the values should be unchanged
+        assert generated_config.log.level == "INFO"
+
+    def test_rename_failure(self, capsys, tmpdir, spec, config_file, old_config):
+        with open(config_file, "w") as f:
+            yaml.safe_dump(
+                old_config, stream=f, default_flow_style=False, encoding="utf-8"
+            )
+
+        with patch("os.rename", Mock(side_effect=ValueError)):
+            generated_config = beer_garden.config.load(
+                ["-c", str(config_file)], force=True
+            )
+
+        # Make sure we printed something
+        assert capsys.readouterr().err
+
+        assert generated_config.log.level == "INFO"
+
+        # Both the tmp file and the old JSON should still be there.
+        assert len(os.listdir(tmpdir)) == 2
+        assert os.path.exists(str(config_file) + ".tmp")
+
+        # The loaded config should be the original file.
+        with open(config_file) as f:
+            new_config_value = yaml.safe_load(f)
+
+        assert new_config_value == old_config
+
+    def test_catastrophe(self, capsys, tmpdir, spec, config_file, old_config):
+        with open(config_file, "w") as f:
+            yaml.safe_dump(
+                old_config, stream=f, default_flow_style=False, encoding="utf-8"
+            )
+
+        with patch("os.rename", Mock(side_effect=[Mock(), ValueError])):
+            with pytest.raises(ValueError):
+                beer_garden.config.load(["-c", str(config_file)], force=True)
+
+        # Make sure we printed something
+        assert capsys.readouterr().err
+
+        # Both the tmp file and the old JSON should still be there.
+        assert len(os.listdir(tmpdir)) == 2
 
 
 @pytest.mark.parametrize(
